add image draw

+++ Src/Core/Image.cpp
@@ -11,7 +11,7 @@ namespace PPEngine {
 
         }
 
-        Image::Image(const std::string& name, int32_t width, int32_t hegiht, int32_t strip, bool alpha, bool hsl, unsigned long mask /*= 0*/)
+        Image::Image(const std::string& name, int32 width, int32 hegiht, int32 strip, bool alpha, bool hsl, unsigned long mask /*= 0*/)
             : name_(name)
             , width_(width)
             , height_(hegiht)
@@ -29,13 +29,13 @@ namespace PPEngine {
         }
 
         Image::Ptr Image::LoadImage(const std::string& name, bool hsl, unsigned long mask) {
-            int32_t width{ 0 }, height{ 0 }, strip{ 0 };
-            uint8_t* image = stbi_load(name.c_str(), &width, &height, &strip, 0);
+            int32 width{ 0 }, height{ 0 }, strip{ 0 };
+            uint8* image = stbi_load(name.c_str(), &width, &height, &strip, 0);
 
             Ptr self = std::make_shared<Image>(name, width, height, strip, strip == 4, hsl, mask);
 
-            uint32_t count = width * height * strip;
-            self->data_ = new uint8_t[count]{ 0 };
+            uint32 count = width * height * strip;
+            self->data_ = new uint8[count]{ 0 };
             self->count_ = count;
             memcpy(self->data_, image, count);
             stbi_image_free(image);
@@ -65,12 +65,12 @@ namespace PPEngine {
             return itor->second;
         }
 
-        const Image::Ptr ImageManager::GetImageEx(const std::string& bitmap, const std::string& type /*= NULL*/, unsigned long mask /*= 0*/, bool bUseHSL /*= false*/) {
+        const Image::Ptr ImageManager::GetImageEx(const std::string& bitmap, const std::string& type /*= NULL*/, uint32 mask /*= 0*/, bool bUseHSL /*= false*/) {
             return nullptr;
         }
 
         const Image::Ptr ImageManager::AddImage(const std::string& bitmap, const std::string& type /*= NULL*/,
-            unsigned long mask /*= 0*/, bool bUseHSL /*= false*/, bool bShared /*= false*/) {
+            uint32 mask /*= 0*/, bool bUseHSL /*= false*/, bool bShared /*= false*/) {
             Image::Ptr image = Image::LoadImage(bitmap, bUseHSL, mask);
             if (bShared) {
                 sharedImages_.insert(std::make_pair(bitmap, image));
